Website cleanup and UI improvements
I've done a bit of editing.

The website will face a more minimalist approach on things to keep things very readable for the user whilst providing the same level of information.

This includes changing the 404 image.

The website has gone a colour change, the background is now darker to help with the other dark items inside of the website.

The live.html page has been updated with the new style, but YT live streaming was not fixable, staying with Restream for now.

Splashtexts on the home screen have been updated and fixed.